- Update to svn r4064
- Added and Update Dosbox Optionals to svn r4064
Allow for direct changing of 4op chaining without having to rewrite the
algorithm mode.

- Added a Missing Resolution Index 0x68, If you set in 'The Big Red
Adventure' S3 Chip
Set and it founded. It saved in the bat the commandline '-24 (For S3)
and the
Resolution Index 104'. this is Index 0x68. On Gamestart it fails because
not founded.

- Added from Freedos Project The memory Manager Programm 'JEMMEX'. I've
searched the
Internet for the Tool. Now its permanent in Dosbox. I hate this
'searchen'